Split common Android logic into base class, to be reused by Emulators/Genymotion/Devices
